Replace custom yaml implementation with go package
This also changes yaml.Unmarshal use to yaml.UnmarshalStrict,
which checks if keys in yaml file are not duplicated.
